Reworks the drawing of the fuel flow gauges
Instead of hardcoding the rate at which the indicator travels around
the fuel flow gauge this commit extracts a constant for the ratio
between throttle and fuel flow. The fuel flow is a guess on my part so
this allows me to change it in the future if more information comes to
light.